Cannot fix the build issues - cancelling this for now.
Detail of issue:
- Can follow the exact steps outlined in this document (cloning the repo, restoring nuget packages and building using msbuild.exe ) locally and it successfully builds.
However, when run through a runner, it fails to build for an unknown reason. The runner is using the exact same executables as above.

In order to prevent time wasting this will need to be continued at a later date. The commit history has also been flooded with unnecessary commits trying to test the issue.